power_status/requirements.txt
PyYAML
emoji
